backend/requirements.txt
aniso8601==8.0.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
oauthlib==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
twython==3.7.0
urllib3==1.25.6
Werkzeug==0.16.0
nltk==3.4.5
pandas==0.25.3
psycopg2-binary==2.8.4
